refactoring detectors
