+++ giskard_vision/core/detectors/perturbation.py
@@ -1,25 +1,15 @@
 import os
 from abc import abstractmethod
+from importlib import import_module
 from pathlib import Path
-from typing import Any, Sequence
+from typing import Any, Sequence, Tuple
 
 import cv2
 
 from giskard_vision.core.dataloaders.wrappers import FilteredDataLoader
-from giskard_vision.core.detectors.base import (
-    DetectorVisionBase,
-    IssueGroup,
-    ScanResult,
-)
+from giskard_vision.core.detectors.base import DetectorVisionBase, ScanResult
+from giskard_vision.core.issues import Robustness
 from giskard_vision.core.tests.base import TestDiffBase
-from giskard_vision.utils.errors import GiskardImportError
-
-from .metrics import detector_metrics
-
-Robustness = IssueGroup(
-    "Robustness",
-    description="Images from the dataset are blurred, recolored and resized to test the robustness of the model to transformations.",
-)
 
 
 class PerturbationBaseDetector(DetectorVisionBase):
@@ -40,6 +30,28 @@ class PerturbationBaseDetector(DetectorVisionBase):
 
     issue_group = Robustness
 
+    def run(
+        self,
+        model: Any,
+        dataset: Any,
+        features: Any | None = None,
+        issue_levels: Tuple[Any] = None,
+        embed: bool = True,
+        num_images: int = 0,
+    ) -> Sequence[Any]:
+        module = import_module(f"giskard_vision.{model.model_type}.detectors.specs")
+        DetectorSpecs = getattr(module, "DetectorSpecs")
+
+        if DetectorSpecs:
+            # Only set attributes that are not part of Python's special attributes (those starting with __)
+            for attr_name, attr_value in vars(DetectorSpecs).items():
+                if not attr_name.startswith("__") and hasattr(self, attr_name):
+                    setattr(self, attr_name, attr_value)
+        else:
+            raise ValueError(f"No detector specifications found for model type: {model.model_type}")
+
+        return super().run(model, dataset, features, issue_levels, embed, num_images)
+
     @abstractmethod
     def get_dataloaders(self, dataset: Any) -> Sequence[Any]: ...
 
@@ -48,7 +60,7 @@ class PerturbationBaseDetector(DetectorVisionBase):
 
         results = []
         for dl in dataloaders:
-            test_result = TestDiffBase(metric=detector_metrics[model.model_type], threshold=1).run(
+            test_result = TestDiffBase(metric=self.metric, threshold=1).run(
                 model=model,
                 dataloader=dl,
                 dataloader_ref=dataset,
@@ -63,40 +75,22 @@ class PerturbationBaseDetector(DetectorVisionBase):
 
             if isinstance(dl, FilteredDataLoader):
                 filename_example_dataloader_ref = str(Path() / "examples_images" / f"{dataset.name}_{index_worst}.png")
-                cv2.imwrite(
-                    filename_example_dataloader_ref, cv2.resize(dataset[index_worst][0][0], (0, 0), fx=0.3, fy=0.3)
-                )
+                cv2.imwrite(filename_example_dataloader_ref, dataset[index_worst][0][0])
                 filename_examples.append(filename_example_dataloader_ref)
 
             filename_example_dataloader = str(Path() / "examples_images" / f"{dl.name}_{index_worst}.png")
-            cv2.imwrite(filename_example_dataloader, cv2.resize(dl[index_worst][0][0], (0, 0), fx=0.3, fy=0.3))
+            cv2.imwrite(filename_example_dataloader, dl[index_worst][0][0])
             filename_examples.append(filename_example_dataloader)
-            results.append(self.get_scan_result(test_result, filename_examples, dl.name, len(dl)))
+            results.append(
+                self.get_scan_result(
+                    test_result.metric_value_test,
+                    test_result.metric_value_test,
+                    test_result.metric_name,
+                    filename_examples,
+                    dl.name,
+                    len(dl),
+                    issue_group=self.issue_group,
+                )
+            )
 
         return results
-
-    def get_scan_result(self, test_result, filename_examples, name, size_data) -> ScanResult:
-        try:
-            from giskard.scanner.issues import IssueLevel
-        except (ImportError, ModuleNotFoundError) as e:
-            raise GiskardImportError(["giskard"]) from e
-
-        relative_delta = (test_result.metric_value_test - test_result.metric_value_ref) / test_result.metric_value_ref
-
-        if relative_delta > self.issue_level_threshold + self.deviation_threshold:
-            issue_level = IssueLevel.MAJOR
-        elif relative_delta > self.issue_level_threshold:
-            issue_level = IssueLevel.MEDIUM
-        else:
-            issue_level = IssueLevel.MINOR
-
-        return ScanResult(
-            name=name,
-            metric_name=test_result.metric_name,
-            metric_value=test_result.metric_value_test,
-            metric_reference_value=test_result.metric_value_ref,
-            issue_level=issue_level,
-            slice_size=size_data,
-            filename_examples=filename_examples,
-            relative_delta=relative_delta,
-        )
