+++ giskard_vision/core/detectors/transformation_color_detector.py
@@ -1,7 +1,7 @@
 from giskard_vision.core.dataloaders.wrappers import ColoredDataLoader
 
 from ...core.detectors.decorator import maybe_detector
-from .perturbation import PerturbationBaseDetector, Robustness
+from .perturbation import PerturbationBaseDetector
 
 
 @maybe_detector("coloring", tags=["vision", "robustness", "image_classification", "landmark", "object_detection"])
@@ -9,8 +9,6 @@ class TransformationColorDetectorLandmark(PerturbationBaseDetector):
     """
     Detector that evaluates models performance depending on images in grayscale
     """
-
-    issue_group = Robustness
 
     def get_dataloaders(self, dataset):
         dl = ColoredDataLoader(dataset)
