+++ giskard_vision/core/detectors/transformation_noise_detector.py
@@ -4,7 +4,7 @@ from ...core.detectors.decorator import maybe_detector
 from .perturbation import PerturbationBaseDetector
 
 
-@maybe_detector("noise", tags=["vision", "robustness", "image_classification", "landmark", "object_detection"])
+@maybe_detector("noise", tags=["vision", "robustness", "image_classification", "landmark", "object_detection", "noise"])
 class TransformationNoiseDetectorLandmark(PerturbationBaseDetector):
     """
     Detector that evaluates models performance on noisy images
