+++ giskard_vision/landmark_detection/detectors/cropping_detector.py
@@ -1,7 +1,9 @@
+from giskard_vision.core.scanner.issues import Robustness
+
 from ...core.detectors.decorator import maybe_detector
 from ..dataloaders.wrappers import CroppedDataLoader
 from ..marks.facial_parts import FacialParts
-from .base import Cropping, LandmarkDetectionBaseDetector
+from .base import LandmarkDetectionBaseDetector
 
 
 @maybe_detector("cropping_landmark", tags=["vision", "face", "landmark", "transformed", "cropped"])
@@ -10,7 +12,7 @@ class CroppingDetectorLandmark(LandmarkDetectionBaseDetector):
     Detector that evaluates models performance relative to a facial part
     """
 
-    issue_group = Cropping
+    issue_group = Robustness
 
     def get_dataloaders(self, dataset):
         facial_parts = [elt.value for elt in FacialParts]
