+++ giskard_vision/landmark_detection/detectors/metadata_detector.py
@@ -3,14 +3,11 @@ from giskard_vision.landmark_detection.detectors.surrogate_functions import (
     SurrogateNME,
     SurrogateVolumeConvexHull,
 )
-from giskard_vision.landmark_detection.tests.performance import NMEMean
 
 from ...core.detectors.decorator import maybe_detector
+from .specs import DetectorSpecs
 
 
 @maybe_detector("metadata_landmark", tags=["vision", "face", "landmark", "metadata"])
-class MetaDataScanDetectorLandmark(MetaDataScanDetector):
+class MetaDataScanDetectorLandmark(DetectorSpecs, MetaDataScanDetector):
     surrogates = [SurrogateVolumeConvexHull, SurrogateNME]
-    metric = NMEMean
-    type_task = "regression"
-    metric_type = "relative"
